C03/requirements.txt
django
mysqlclient
gunicorn
pyjwt
PyMySQL
scrypt
python-dotenv
pyyaml
djangorestframework
Pillow
django_filter
django-cors-headers
django-crontab
django-apscheduler
apscheduler